+++ wizards/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
 from . import partner_statement_wizard
+from . import whatsapp_statement_wizard